pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "sfn2024-behavior-hw-demo"
description = "A repository for demos of hardware used by the behavior team."
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3"
]
readme = "README.md"
dynamic = ["version"]

dependencies = [
  "aind_behavior_services@git+https://github.com/AllenNeuralDynamics/Aind.Behavior.Services@0.8.0-rc2",
]

[project.optional-dependencies]

linters = [
    'ruff',
    'codespell'
]

docs = [
    'Sphinx<7.3',
    'furo',
    'sphinx-jinja',
    'autodoc_pydantic[erdantic]',
    'myst_parser',
    'sphinx-jsonschema'
]

[tool.setuptools.packages.find]
where = ["src/DataSchemas"]

[tool.setuptools.dynamic]
version = {attr = "sfn2024_behavior_hw_demo.__version__"}
readme = {file = ["README.md"]}

[tool.ruff]
line-length = 120
target-version = 'py311'

[tool.ruff.lint]
extend-select = ['Q', 'RUF100', 'C90', 'I']
extend-ignore = []
mccabe = { max-complexity = 14 }
pydocstyle = { convention = 'google' }

[tool.codespell]
skip = '.git,*.pdf,*.svg,./bonsai,*.bonsai,./docs/_build'
ignore-words-list = 'nd'